Initial commit for remediations
A remediation is an action to be accomplished by a user in order to fix
a vulnerability.

A single remediation could fix a set of vulnerabilities, and a single
vulnerability could be fix by a set of remediations, therefore they have
a many-to-many relationship.

The URLs for create, edit and associate a remediation with a
vulnerability are:
remediations/add.html
remediations/edit.html/(id)
remediations/associate_vuln.html

The host's detail page will list the remediations applicable for the
hosts, based on the vulnerabilities affecting the host.

In the same page the user can hit the 'Apply' button on the remediation
to inform that the action was taken on the host. Comments and
attachments could be used to follow the state of the applied
remediation.